BTE_APP_backend/requirements.txt
fastapi
sqlalchemy
pymysql
uvicorn
pytz
mangum
pydantic==1.10.13
pydantic-core
boto3
python-dotenv
azure-functions
nest-asyncio
azure-identity
azure-keyvault-secrets
